Initial input rendering
